pyproject.toml
[tool.poetry]
name = "hanghae-hackathon"
version = "0.1.0"
description = ""
authors = ["eunsung <edplab2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.26.1"
uvicorn = "^0.32.0"
fastapi = "^0.115.5"
langchain-community = "^0.3.7"
python-dotenv = "^1.0.1"
openai = "^1.54.4"
langchain-chroma = "^0.1.4"
langchain = "^0.3.7"
langchain-core = "^0.3.18"
gspread = "^6.1.4"
toml = "^0.10.2"
oauth2client = "^4.1.3"
langchain-openai = "^0.2.8"
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pypdf = "^5.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
